+++ api/gl/Canvas.h
@@ -67,7 +67,7 @@ namespace v3d::gl {
         // vertex data
         std::vector<glm::vec3> xyz_;
         std::vector<glm::vec4> rgba_;
-        std::vector<size_t> indices_;
+        std::vector<unsigned int> indices_;
     };
 
 };  // namespace v3d::gl
